Move tests to correct module
