+++ sweetest/src/test/java/com/mysugr/sweetest/coroutine/FlowTest.kt
@@ -1,7 +1,7 @@
-package com.mysugr.android.testing.example.coroutine
+package com.mysugr.sweetest.coroutine
 
-import com.mysugr.android.testing.example.appModuleTestingConfiguration
 import com.mysugr.sweetest.framework.base.BaseJUnitTest
+import com.mysugr.sweetest.framework.configuration.moduleTestingConfiguration
 import com.mysugr.sweetest.framework.coroutine.runBlockingSweetest
 import com.mysugr.sweetest.framework.coroutine.testCoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -10,16 +10,29 @@ import kotlinx.coroutines.channels.BroadcastChannel
 import kotlinx.coroutines.flow.asFlow
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.test.UncompletedCoroutinesError
 import org.junit.Assert.assertEquals
+import org.junit.Test
 
 @ExperimentalCoroutinesApi
 @FlowPreview
-class FlowTestWithAutoCancel : BaseJUnitTest(appModuleTestingConfiguration) {
+class FlowTest : BaseJUnitTest(moduleTestingConfiguration()) {
 
-    override fun configure() = super.configure()
-        .autoCancelTestCoroutines(true) // <-- !
+    @Test
+    fun `Hot Flow with TestCoroutineScope is successful`() {
+        val channel = BroadcastChannel<Int>(1)
+        val emissions = mutableListOf<Int>()
 
-    fun `In contrast to FlowTest without auto cancel, here no exception is expected`() = runBlockingSweetest {
+        channel.asFlow()
+            .onEach { emissions.add(it) }
+            .launchIn(testCoroutineScope)
+        channel.offer(1)
+
+        assertEquals(1, emissions.size)
+    }
+
+    @Test(expected = UncompletedCoroutinesError::class)
+    fun `Hot Flow produces error when not finished`() = runBlockingSweetest {
         val channel = BroadcastChannel<Int>(1)
         val emissions = mutableListOf<Int>()
 
@@ -29,7 +42,5 @@ class FlowTestWithAutoCancel : BaseJUnitTest(appModuleTestingConfiguration) {
 
         channel.offer(1)
         assertEquals(1, emissions.size)
-
-        // After the test sweetest automatically cancels the job created by `launchIn`
     }
 }
